I put the sentences loading one layer outside the main editor, to avoid the editor to re-render everytime the juxtalinear editor module
